+++ omni/benchmark/validation/validator.py
@@ -1,13 +1,15 @@
+import logging
 import os.path
 from collections import Counter
 from pathlib import Path
 from typing import Union, Optional
 from urllib.parse import urlparse
 
-from omni_schema.datamodel.omni_schema import SoftwareBackendEnum
+from omni_schema.datamodel.omni_schema import SoftwareBackendEnum, SoftwareEnvironment
 
 from omni.benchmark.converter import LinkMLConverter
 from omni.benchmark.validation.error import ValidationError
+from omni.utils import try_avail_envmodule
 
 
 class Validator:
@@ -71,49 +73,42 @@ class Validator:
 
         software_backend = converter.get_software_backend()
         for environment in software_environments.values():
-            environment_exists = True
-            environment_path = None
-            if (
-                software_backend == SoftwareBackendEnum.apptainer
+            if software_backend == SoftwareBackendEnum.host:
+                continue
+
+            elif software_backend == SoftwareBackendEnum.envmodules:
+                load_result = try_avail_envmodule(environment.envmodule)
+                if not load_result:
+                    self.errors.append(
+                        ValidationError(
+                            f"Software environment with id '{environment.id}' and name '{environment.envmodule}' could not be loaded as a valid `envmodule`."
+                        )
+                    )
+            elif (
+                software_backend == SoftwareBackendEnum.conda
                 or software_backend == SoftwareBackendEnum.docker
+                or software_backend == SoftwareBackendEnum.apptainer
             ):
                 environment_path = Validator.get_environment_path(
-                    benchmark_dir, environment.apptainer
+                    software_backend, environment, benchmark_dir
                 )
 
-                if environment.apptainer is None:
-                    environment_exists = False
-
-            if software_backend == SoftwareBackendEnum.envmodules:
-                environment_path = Validator.get_environment_path(
-                    benchmark_dir, environment.easyconfig
-                )
-
-                if environment.envmodule is None:
-                    environment_exists = False
-
-            if software_backend == SoftwareBackendEnum.conda:
-                environment_path = Validator.get_environment_path(
-                    benchmark_dir, environment.conda
-                )
-
-                if environment.conda is None:
-                    environment_exists = False
-
-            if not environment_exists:
-                self.errors.append(
-                    ValidationError(
-                        f"Software environment with id '{environment.id}' does not define the following backend: '{software_backend.text}'."
+                if not environment_path:
+                    self.errors.append(
+                        ValidationError(
+                            f"Software environment with id '{environment.id}' does not have a valid backend definition for: '{software_backend.text}'."
+                        )
                     )
-                )
-            elif not Validator.is_url(environment_path) and not os.path.exists(
-                environment_path
-            ):
-                self.errors.append(
-                    ValidationError(
-                        f"Software environment path for '{software_backend.text}' does not exist: '{environment_path}'."
+
+                # else for other backends, attempt to find path on disk
+                elif not Validator.is_url(environment_path) and not os.path.exists(
+                    environment_path
+                ):
+                    self.errors.append(
+                        ValidationError(
+                            f"Software environment path for '{software_backend.text}' does not exist: '{environment_path}'."
+                        )
                     )
-                )
 
         # Raise ValidationError if there are errors
         if self.errors:
@@ -130,12 +125,29 @@ class Validator:
 
     @staticmethod
     def get_environment_path(
-        benchmark_dir: Path, environment: Optional[str]
+        software_backend: SoftwareBackendEnum,
+        software: SoftwareEnvironment,
+        benchmark_dir: Path,
     ) -> Optional[str]:
+        environment = None
+        if (
+            software_backend == SoftwareBackendEnum.apptainer
+            or software_backend == SoftwareBackendEnum.docker
+        ):
+            environment = software.apptainer
+
+        elif software_backend == SoftwareBackendEnum.conda:
+            environment = software.conda
+
+        elif software_backend == SoftwareBackendEnum.envmodules:
+            environment = software.envmodule
+
         if not environment:
             return None
 
         if Validator.is_url(environment) or Validator.is_absolute_path(environment):
+            environment_path = environment
+        elif software_backend == SoftwareBackendEnum.envmodules:
             environment_path = environment
         else:
             environment_path = os.path.join(benchmark_dir, environment)